fix: cache Flutter Engine
The main reason for having a singleton is o cache Flutter Engine.

By decoupling the creation of the Flutter Engine we can run its initialization
synchronously in the main thread when TikiPlatformChannel is loaded.

In subsequent TikiSdk.initSdk calls, we just need to update the blockchain,
keeping the same Flutter Engine.
Changes
	modified:   IntegrationTests/IntegrationTests/IntegrationTests.swift
	modified:   Sources/TikiSdk/TikiPlatformChannel/TikiPlatformChannel.swift
	modified:   Sources/TikiSdk/TikiSdk.swift